requirements.txt
numpy>=1.19.2
torch>=1.7.0
python-chess>=1.999
matplotlib>=3.3.2
tqdm>=4.50.2
pandas>=1.1.3
tensorboard>=2.3.0 